getStimvol.m, getGlmEVParamsGUI.m, makeDesignMatrix.m:Added new GLM analysis parameter acquisitionDuration.
Previously, when downsampling a supersampled design matrix, it was assumed that the
acquisition duration was equal to the frame period, (the design matrix was dowsampled by integrating, downsampling and differentiating, which is equivalent to averaging subsamples corresponding to a full frame period around the acquisition time). However this doesn't correspond very well to some situations, like sparse sampling. The ne acquisitionDuration parameter (in seconds) indicates how many subsamples should be integrated when downsampling the design matrix. By default, it is set to the frame period.
Details of changes:
getStimvol.m: if the supersamplingMode parameter is set to 'automatic',
acquisitionDuration is added to the other stimulus onsets and durations
that are used to compute the designSupersampling value.
getGlmEVParamsGUI.m: added parameter to GUI and default value.
makeDesignMatrix.m: replace call to mrDownsample by custom downsampling
method that convert the acquisitionDuration value from seconds to
subsamples and integrates the required subsamples into each new sample.